cloudwatch: Move tests to main_test.go
